feat(students) create DELETE endpoint and some fix on FindOne endpoint

+++ src/students/students.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { StudentsRepository } from "./students.repository";
 import { FailToCreateError } from "./errors/FailToCreate.error";
 import { EStudentsErrors } from "./errors/types/studentsErrors";
+import { NotFoundError } from "./errors/NotFound.error";
 
 @Injectable()
 export class StudentsService {
@@ -26,7 +27,7 @@ export class StudentsService {
   ): Promise<string> {
     const student = await this.studentsRepository.findOneByCPF(cpf);
 
-    if (Object.keys(student).length === 0) {
+    if (Object.keys(student).length < 0) {
       throw new FailToCreateError(EStudentsErrors.STUDENT_EXISTS);
     }
 
@@ -45,7 +46,15 @@ export class StudentsService {
     return data;
   }
 
-  public delete(id: string) {
-    return this.studentsRepository.delete(id);
+  public async delete(id: string) {
+    const student = await this.studentsRepository.findOne(id);
+
+    if (Object.keys(student).length === 0) {
+      return {};
+    }
+
+    await this.studentsRepository.delete(id);
+
+    return student.id;
   }
 }
